+++ src/tsfile/encoding/encoder/rleEncoder.go
@@ -1,3 +1,22 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 package encoder
 
 import (
@@ -15,9 +34,9 @@ type RleEncoder struct {
 	encodeEndian        int16
 	values_32           []int32
 	values_64           []int64
-	bitWidth            int32
-	repeatCount         int32
-	bitPackedGroupCount int32
+	bitWidth            int
+	repeatCount         int
+	bitPackedGroupCount int
 	bytesBuffer         []byte
 	isBitPackRun        bool
 	preValue_32         int32
@@ -26,7 +45,7 @@ type RleEncoder struct {
 	byteCache           *bytes.Buffer
 	packer_32           *bitpacking.IntPacker
 	packer_64           *bitpacking.LongPacker
-	numBufferedValues   int32
+	numBufferedValues   int
 	bufferedValues_32   []int32
 	bufferedValues_64   []int64
 }
@@ -71,7 +90,7 @@ func (this *RleEncoder) EncInt64(value int64, buffer *bytes.Buffer) {
 	this.values_64 = append(this.values_64, value)
 }
 
-func getIntMaxBitWidth(list []int32) int32 {
+func getIntMaxBitWidth(list []int32) int {
 	max := int32(1)
 	for _, num := range list {
 		bitWidth := 32 - utils.NumberOfLeadingZeros(num)
@@ -79,10 +98,10 @@ func getIntMaxBitWidth(list []int32) int32 {
 			max = bitWidth
 		}
 	}
-	return max
+	return int(max)
 }
 
-func getLongMaxBitWidth(list []int64) int32 {
+func getLongMaxBitWidth(list []int64) int {
 	max := int32(1)
 	for _, num := range list {
 		bitWidth := 64 - utils.NumberOfLeadingZerosLong(num)
@@ -90,7 +109,7 @@ func getLongMaxBitWidth(list []int64) int32 {
 			max = bitWidth
 		}
 	}
-	return max
+	return int(max)
 }
 
 func (this *RleEncoder) endPreviousBitPackedRun(lastBitPackedNum int32) {
@@ -132,8 +151,8 @@ func (this *RleEncoder) convertBuffer() {
 	case (constant.BOOLEAN):
 	case (constant.INT32):
 		tmpBuffer := make([]int32, conf.RLE_MIN_REPEATED_NUM)
-		for i := int32(0); i < conf.RLE_MIN_REPEATED_NUM; i++ {
-			if i < int32(len(this.bufferedValues_32)) {
+		for i := 0; i < conf.RLE_MIN_REPEATED_NUM; i++ {
+			if i < len(this.bufferedValues_32) {
 				tmpBuffer[i] = (this.bufferedValues_32[i])
 			}
 		}
@@ -142,8 +161,8 @@ func (this *RleEncoder) convertBuffer() {
 		break
 	case (constant.INT64):
 		tmpBuffer := make([]int64, 0) //conf.RLE_MIN_REPEATED_NUM)
-		for i := int32(0); i < conf.RLE_MIN_REPEATED_NUM; i++ {
-			if i < int32(len(this.bufferedValues_64)) {
+		for i := 0; i < conf.RLE_MIN_REPEATED_NUM; i++ {
+			if i < len(this.bufferedValues_64) {
 				tmpBuffer = append(tmpBuffer, this.bufferedValues_64[i])
 			} else {
 				tmpBuffer = append(tmpBuffer, 0)
@@ -229,12 +248,12 @@ func (this *RleEncoder) clearBuffer() {
 		switch this.tsDataType {
 		case (constant.BOOLEAN):
 		case (constant.INT32):
-			if i < int32(len(this.bufferedValues_32)) {
+			if i < len(this.bufferedValues_32) {
 				this.bufferedValues_32 = append(this.bufferedValues_32, 0)
 			}
 			break
 		case (constant.INT64):
-			if i < int32(len(this.bufferedValues_64)) {
+			if i < len(this.bufferedValues_64) {
 				this.bufferedValues_64 = append(this.bufferedValues_64, 0)
 			}
 			break
@@ -289,15 +308,15 @@ func (this *RleEncoder) Flush(buffer *bytes.Buffer) {
 	switch this.tsDataType {
 	case (constant.BOOLEAN):
 	case (constant.INT32):
-		this.bitWidth = getIntMaxBitWidth(this.values_32)
-		this.packer_32 = &bitpacking.IntPacker{BitWidth: this.bitWidth}
+		this.bitWidth = int(getIntMaxBitWidth(this.values_32))
+		this.packer_32 = &bitpacking.IntPacker{BitWidth: int(this.bitWidth)}
 		for _, v := range this.values_32 {
 			this.encodeValue(int64(v))
 		}
 		break
 	case (constant.INT64):
-		this.bitWidth = getLongMaxBitWidth(this.values_64)
-		this.packer_64 = &bitpacking.LongPacker{BitWidth: this.bitWidth}
+		this.bitWidth = int(getLongMaxBitWidth(this.values_64))
+		this.packer_64 = &bitpacking.LongPacker{BitWidth: int(this.bitWidth)}
 		for _, v := range this.values_64 {
 			this.encodeValue(int64(v))
 		}
